+++ simpleble/src/frontends/base/Advanced.cpp
@@ -1,26 +1,34 @@
 #include "simpleble/Advanced.h"
 
 #if defined(_WIN32)
-    // Windows code
+namespace SimpleBLE::Advanced::Windows {}
 
 #endif
 
 #if TARGET_OS_OSX
-    // macOS code
+namespace SimpleBLE::Advanced::MacOS {}
 
 #endif
 
 #if TARGET_OS_IOS
-    // iOS code
+namespace SimpleBLE::Advanced::iOS {}
 
 #endif
 
 #if defined(__ANDROID__)
-    // Android code
+
+#include "simplejni/VM.hpp"
+
+namespace SimpleBLE::Advanced::Android {
+
+JavaVM* get_jvm() { return SimpleJNI::VM::jvm(); }
+void set_jvm(JavaVM* jvm) { SimpleJNI::VM::jvm(jvm); }
+
+}  // namespace SimpleBLE::Advanced::Android
 
 #endif
 
 #if defined(__linux__) && !defined(__ANDROID__)
-    // Linux code
+namespace SimpleBLE::Advanced::Linux {}
 
 #endif